chore(clean-types): renames type_wip -> types
