+++ runem/report.py
@@ -4,8 +4,8 @@ from collections import defaultdict
 from datetime import timedelta
 
 from runem.log import log
-from runem.types_wip.common import OrderedPhases, PhaseName
-from runem.types_wip.types_jobs import (
+from runem.types.common import OrderedPhases, PhaseName
+from runem.types.types_jobs import (
     JobReturn,
     JobRunMetadatasByPhase,
     JobRunReportByPhase,
