+++ runem/types/runem_config.py
@@ -1,9 +1,9 @@
 import pathlib
 import typing
 
-from runem.types_wip.common import JobName, JobTags, OrderedPhases, PhaseName
-from runem.types_wip.filters import TagFileFilter
-from runem.types_wip.hooks import HookName
+from runem.types.common import JobName, JobTags, OrderedPhases, PhaseName
+from runem.types.filters import TagFileFilter
+from runem.types.hooks import HookName
 
 
 class OptionConfig(typing.TypedDict, total=False):
