+++ runem/types_public_api.py
@@ -1,6 +1,6 @@
-from runem.types_wip.common import FilePathList, JobName
-from runem.types_wip.options import Options
-from runem.types_wip.types_jobs import HookKwargs, JobKwargs, JobReturnData
+from runem.types.common import FilePathList, JobName
+from runem.types.options import Options
+from runem.types.types_jobs import HookKwargs, JobKwargs, JobReturnData
 
 __all__ = [
     "FilePathList",
