+++ tests/test_job.py
@@ -4,9 +4,9 @@ from collections import defaultdict
 import pytest
 
 from runem.job import Job, NoJobName
-from runem.types_wip.common import FilePathList, JobTags
-from runem.types_wip.filters import FilePathListLookup
-from runem.types_wip.runem_config import JobConfig
+from runem.types.common import FilePathList, JobTags
+from runem.types.filters import FilePathListLookup
+from runem.types.runem_config import JobConfig
 
 
 @pytest.mark.parametrize(
