+++ tests/test_runem.py
@@ -27,9 +27,9 @@ from runem.runem import (
     _update_progress,
     timed_main,
 )
-from runem.types_wip.filters import FilePathListLookup
-from runem.types_wip.hooks import HookName
-from runem.types_wip.runem_config import (
+from runem.types.filters import FilePathListLookup
+from runem.types.hooks import HookName
+from runem.types.runem_config import (
     Config,
     GlobalSerialisedConfig,
     HookSerialisedConfig,
@@ -39,7 +39,7 @@ from runem.types_wip.runem_config import (
     PhaseGroupedJobs,
     UserConfigMetadata,
 )
-from runem.types_wip.types_jobs import JobReturn, JobTiming
+from runem.types.types_jobs import JobReturn, JobTiming
 from tests.intentional_test_error import IntentionalTestError
 from tests.sanitise_reports_footer import sanitise_reports_footer
